Move the localnotifications package out of the login folder
